Task 1
This To-Do List app will be a simple, single-page application with a clean and intuitive user interface. The front-end will consist of a form allowing users to input a task name and description. Upon submission, the task will be added to a dynamically updating list displayed below. Each task in the list will include checkboxes for marking completion and buttons for deletion.